+++ lib/OpcodeTrait.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * Copyright (C) 2014-2021 Textalk/Abicart and contributors.
+ *
+ * This file is part of Websocket PHP and is free software under the ISC License.
+ * License text: https://raw.githubusercontent.com/Textalk/websocket-php/master/COPYING
+ */
+
+namespace WebSocket;
+
+trait OpcodeTrait
+{
+    private static $opcodes = [
+        'continuation' => 0,
+        'text'         => 1,
+        'binary'       => 2,
+        'close'        => 8,
+        'ping'         => 9,
+        'pong'         => 10,
+    ];
+}
